Entities moved to common/models
For now we keep all the entities in one place
